requirements.txt
aioboto3==13.4.0
aiobotocore==2.18.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.12
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
boto3==1.36.1
botocore==1.36.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
decorator==5.2.1
donoratlas @ file:///home/ec2-user/DonorAtlas/DonorAtlas
frozenlist==1.7.0
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geonamescache==2.0.0
geopy==2.4.1
haversine==2.9.0
humanize==4.12.2
idna==3.10
jmespath==1.0.1
joblib==1.5.1
libarchive-c==5.3
lxml==6.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.5.1
nameparser==1.1.3
nicknames==0.1.11
nltk==3.9.1
numpy==2.3.1
orjson==3.10.18
pandas==2.2.3
pdfrw==0.4
polars==1.31.0
probableparsing==0.0.1
propcache==0.3.2
psutil==6.1.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.2
PyPDF2==3.0.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
ratelim==0.1.6
regex==2024.11.6
requests==2.32.3
rich==14.0.0
s3transfer==0.11.3
simstring-pure==1.0.0
six==1.17.0
soupsieve==2.7
tabulate==0.9.0
termcolor==3.0.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
Unidecode==1.3.8
urllib3==2.5.0
usaddress==0.5.14
usaddress-scourgify==0.6.0
wrapt==1.17.2
xxhash==3.5.0
yaml-config==0.1.5
yarl==1.20.1

